requirements.txt
fastapi
uvicorn
web3
python-dotenv
pydantic
pymongo
requests